feat(faultdetection): added fault detection benchmarking scripts;
This is a first version of benchmarking code that looks at the prometheus response and utilises the experiment report to evaluate if the injected faults were discovered, and if so, how fast.
I focused on producing a terminal output for now, no "scoring" implemented.
Additionally, there are some alerts that I chose to ignore, because they clutter the output a lot (they are constantly firing).